pyproject.toml
[project]
name = "swarm_construction"
version = "0.1.0"
description = ""
authors = [
    {name = "Tom Milner",email = "el20tm@leeds.ac.uk"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy (>=2.2.4,<3.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "tk (>=0.1.0,<0.2.0)",
    "snakeviz (>=2.2.2,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[tool.poetry]
package-mode = false